メンション関係のテストを書き換えた
